rename and improve warn messages
