+++ lint/testdata/uselessLen/negative_tests.go
